Improve accessibility of the language picker
Add a button that toggles its 'aria-expanded' attribute between 'true'
and 'false' which shows or hides the available languages accordingly.

When the language-picker is open a keydown handler is bound to the
document that will close the language-picker when Escape is pressed. The
same goes for a click handler that watches for clicks outside the
language-picker itself.