Addition of Subject to Mailto Links
By appending the subject line, we ease the user from having to think about the subject line and can focus on the message.

If unchanged, it also gives a clear indication to the recepient of the email on the context of the email message from the subject itself.

A rule based system could also be created for such emails to be forwarded as required, or placed in a particular folder, helping the recepient who has to process the Corona Melder inbox.

Used the translations.yml to set the EN and NL translation strings, so that they can be centrally modified and added as needed.