Addition of Target _Blank to the Mailto Link
This helps users whose email mailto links opens in the browser, so that they retain their previous page just in case they need to refer to contents from the page.

There will also not be the back button, and avoid from having the user use it and losing his message.